# apiserver/swagger/console_swagger.yaml
...
swagger: "2.0"
info:
  description: "The API for the EatBacon IOT project"
  version: "1.0.1"
  title: "home-iot-api"
host: "localhost:4145"
basePath: "/iot/api/v1/console"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /products:
    post:
      tags:
      - "Product"
      operationId: "createProduct"
      parameters:
      - in: "body"
        name: "product"
        required: false
        schema:
          $ref: "#/definitions/productAddRequest"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/ProductResponse"
        400:
          description: "The body of the request is not valid"
          schema:
            $ref: "#/definitions/ApiResponse"
        401:
          description: "The authorization token cannot be validated"
          schema:
            $ref: "#/definitions/ApiResponse"
    patch:
      tags:
      - "Product"
      operationId: "updateProduct"
      parameters:
      - in: "body"
        name: "product"
        required: false
        schema:
          $ref: "#/definitions/productUpdateRequest"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "The body of the request is not valid"
          schema:
            $ref: "#/definitions/ApiResponse"
        401:
          description: "The authorization token cannot be validated"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Product"
  /products/{productId}:
    get:
      tags:
      - "Product"
      operationId: "getProduct"
      parameters:
      - name: "productId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/ProductResponse"
        401:
          description: "The authorization token cannot be validated"
          schema:
            $ref: "#/definitions/ApiResponse"
        404:
          description: "The IoT hub instance or a device identity not exist"
          schema:
            $ref: "#/definitions/ApiResponse"
    delete:
      tags:
      - "Product"
      operationId: "removeProduct"
      parameters:
      - name: "productId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/ApiResponse"
        401:
          description: "The authorization token cannot be validated"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Product"
  /products/{productId}/devices:
    get:
      tags:
      - "Product"
      operationId: "getProductDevices"
      parameters:
      - name: "productId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/ProductDevicesResponse"
        401:
          description: "The authorization token cannot be validated"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Product"
...
  /devices:
    post:
      tags:
      - "Device"
      operationId: "createDevice"
      parameters:
      - in: "body"
        name: "device"
        required: false
        schema:
          $ref: "#/definitions/Device"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/DeviceResponse"
        400:
          description: "The body of the request is not valid"
          schema:
            $ref: "#/definitions/ApiResponse"
        401:
          description: "The authorization token cannot be validated"
          schema:
            $ref: "#/definitions/ApiResponse"
    patch:
      tags:
      - "Device"
      operationId: "updateDevice"
      parameters:
      - in: "body"
        name: "device"
        required: false
        schema:
          $ref: "#/definitions/Device"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "The body of the request is not valid"
          schema:
            $ref: "#/definitions/ApiResponse"
        401:
          description: "The authorization token cannot be validated"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Device"
  /products/{productId}/devices/{deviceId}:
    get:
      tags:
      - "Device"
      operationId: "getDevice"
      parameters:
      - name: "productId"
        in: "path"
        required: true
        type: "string"
      - name: "deviceId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "The body of the request is not valid"
          schema:
            $ref: "#/definitions/ApiResponse"
        401:
          description: "The authorization token cannot be validated"
          schema:
            $ref: "#/definitions/ApiResponse"
    delete:
      tags:
      - "Device"
      operationId: "deleteDevice"
      parameters:
      - name: "productId"
        in: "path"
        required: true
        type: "string"
      - name: "deviceId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "The body of the request is not valid"
          schema:
            $ref: "#/definitions/ApiResponse"
        401:
          description: "The authorization token cannot be validated"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Device"
  /devices/bulk:
    post:
      tags:
      - "Device"
      operationId: "bulkRegisterDevices"
      parameters:
      - in: "body"
        name: "device"
        required: false
        schema:
          $ref: "#/definitions/BulkDevice"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/DeviceResponse"
        400:
          description: "The body of the request is not valid"
          schema:
            $ref: "#/definitions/ApiResponse"
        401:
          description: "The authorization token cannot be validated"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Device"
  /products/{productId}/devices/{deviceId}/shadow:
    get:
      tags:
      - "Device"
      operationId: "getShadowDevice"
      parameters:
      - name: "productId"
        in: "path"
        required: true
        type: "string"
      - name: "deviceId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/DeviceResponse"
        400:
          description: "The body of the request is not valid"
          schema:
            $ref: "#/definitions/ApiResponse"
        401:
          description: "The authorization token cannot be validated"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Device"
  /devices/shadow:
    patch:
      tags:
      - "Device"
      operationId: "updateShadowDevice"
      parameters:
      - in: "body"
        name: "device"
        required: false
        schema:
          $ref: "#/definitions/BulkDevice"
      responses:
        200:
          description: "response"
          schema:
            $ref: "#/definitions/DeviceResponse"
        400:
          description: "The body of the request is not valid"
          schema:
            $ref: "#/definitions/ApiResponse"
        401:
          description: "The authorization token cannot be validated"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "Device"
...
definitions:
...
  ApiResponse:
    type: "object"
    properties:
      RequestId:
        type: "string"
        format: "int32"
      Message:
        type: "string"
        example: "everything is ok"
    example:
      RequestId: "RequestId"
      Message: "everything is ok"
...
  productAddRequest:
    type: "object"
    properties:
      ProductName:
        type: "string"
      Category:
        type: "string"
      Description:
        type: "string"
  productUpdateRequest:
    type: "object"
    properties:
      ProductId:
        type: "string"
      ProductName:
        type: "string"
      Category:
        type: "string"
      Description:
        type: "string"
...
  ProductResponse:
    type: "object"
    properties:
      requestId:
        type: "string"
        example: "fd385079-1db6-420a-bf85-4c2efdc3e19b"
      result:
        type: "object"
        properties: {}
    example:
      result: "{}"
      requestId: "fd385079-1db6-420a-bf85-4c2efdc3e19b"
  ProductDevicesResponse:
    type: "object"
    properties:
      RequestId:
        type: "string"
        example: "fd385079-1db6-420a-bf85-4c2efdc3e19b"
      Success:
        type: "string"
        example: "true"
      Result:
        type: "array"
        items:
          type: "object"
          properties: {}
    example:
      RequestId: "fd385079-1db6-420a-bf85-4c2efdc3e19b"
      Success: "true"
      Result:
      - "{}"
      - "{}"
  Device:
    type: "object"
    properties:
      ProductId:
        type: "string"
      DeviceId:
        type: "string"
      DeviceName:
        type: "string"
      DeviceStatus:
        type: "string"
      DeviceSecret:
        type: "string"
      TimeCreated:
        type: "string"
      TimeModified:
        type: "string"
    example:
      TimeCreated: "TimeCreated"
      TimeModified: "TimeModified"
      DeviceId: "DeviceId"
      DeviceSecret: "DeviceSecret"
      ProductId: "ProductId"
      DeviceStatus: "DeviceStatus"
      DeviceName: "DeviceName"
  BulkDevice:
    type: "object"
    properties:
      Device:
        type: "object"
        properties: {}
      Number:
        type: "string"
    example:
      Number: "Number"
      Device: "{}"
  DeviceResponse:
    type: "object"
    properties:
      requestId:
        type: "string"
        example: "fd385079-1db6-420a-bf85-4c2efdc3e19b"
      result:
        type: "object"
        properties: {}
    example:
      result: "{}"
      requestId: "fd385079-1db6-420a-bf85-4c2efdc3e19b"
...